feat: implement /predict route to get gold or deposit prediction
